requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.2
django-cleanup==6.0.0
django-cloudinary-storage==0.3.0
django-tinymce==3.4.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
